--- src/main/java/ru/clevertec/animal/proxy/annotation/Delete.java
@@ -1,14 +1,0 @@
-package ru.clevertec.animal.proxy.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Аннотация для использования aspectJ для указания операции удаления
- * */
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Delete {
-}
