--- src/main/java/ru/clevertec/animal/proxy/cache/IBaseCache.java
@@ -1,13 +1,0 @@
-package ru.clevertec.animal.proxy.cache;
-
-import java.util.Optional;
-
-public interface IBaseCache<K, V> {
-
-    Optional<V> get(K key);
-
-    void put(K key, V value);
-
-    void remove(K key);
-
-}
