+++ src/main/java/ru/clevertec/proxy/annotation/Put.java
@@ -1,0 +1,14 @@
+package ru.clevertec.proxy.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Аннотация для использования aspectJ для указания операции вставки
+ * */
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Put {
+}
