+++ src/main/java/ru/clevertec/service/IBaseService.java
@@ -1,0 +1,14 @@
+package ru.clevertec.service;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface IBaseService<T> {
+
+    T get(UUID uuid);
+    List<T> getAll();
+    UUID create(T animalDto);
+    void update(UUID uuid, T animalDto);
+    void delete(UUID uuid);
+
+}
