+++ src/test/java/ru/clevertec/animal/service/impl/AnimalServiceImplTest.java
@@ -6,12 +6,13 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import ru.clevertec.animal.dao.IBaseDao;
-import ru.clevertec.animal.data.AnimalDto;
-import ru.clevertec.animal.entity.Animal;
-import ru.clevertec.animal.exception.AnimalNotFoundException;
-import ru.clevertec.animal.exception.ValidatorException;
-import ru.clevertec.animal.mapper.AnimalMapper;
+import ru.clevertec.dao.IBaseDao;
+import ru.clevertec.data.AnimalDto;
+import ru.clevertec.entity.Animal;
+import ru.clevertec.exception.AnimalNotFoundException;
+import ru.clevertec.exception.ValidatorException;
+import ru.clevertec.mapper.AnimalMapper;
+import ru.clevertec.service.impl.AnimalServiceImpl;
 import ru.clevertec.util.AnimalTestData;
 import ru.clevertec.util.AnimalTestDataForService;
