+++ src/test/java/ru/clevertec/util/AnimalTestDataForService.java
@@ -2,9 +2,9 @@ package ru.clevertec.util;
 
 import lombok.Builder;
 import lombok.Data;
-import ru.clevertec.animal.connectionDB.MyConnection;
-import ru.clevertec.animal.data.AnimalDto;
-import ru.clevertec.animal.entity.Animal;
+import ru.clevertec.connectionDB.MyConnection;
+import ru.clevertec.data.AnimalDto;
+import ru.clevertec.entity.Animal;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
